update rename geometry
