+++ src/connectors/normal.mjs
@@ -1,4 +1,4 @@
-import * as g from '../geometry.mjs';
+import * as g from '../g/index.mjs';
 
 export const normal = function(sourcePoint, targetPoint, route, opt) {
