+++ src/dia/Element.mjs
@@ -1,5 +1,5 @@
 import { Cell } from './Cell.mjs';
-import { Point, toRad, normalizeAngle, Rect } from '../geometry.mjs';
+import { Point, toRad, normalizeAngle, Rect } from '../g/index.mjs';
 import { isNumber, isObject, interpolate, assign, invoke, normalizeSides } from '../util/util.js';
 import { elementPortPrototype } from './ports.js';
