+++ src/dia/Graph.mjs
@@ -1,6 +1,6 @@
 import Backbone from 'backbone';
 import * as util from '../util/util.js';
-import * as g from '../geometry.mjs';
+import * as g from '../g/index.mjs';
 
 import { Link } from './Link.mjs';
 import { Element } from './Element.mjs';
