+++ src/layout/index.mjs
@@ -1,5 +1,5 @@
 import * as Port from './ports/joint.layout.port.mjs'
-import * as PortLabel from './portsportLabel.mjs'
+import * as PortLabel from './ports/portLabel.mjs'
 
 export { Port, PortLabel }
 export * from './DirectedGraph/DirectedGraph.mjs';
